Build the C++ chains again after the move to compute
Three things broke when the D and F ports and the GPU code left this
repository for compute, and none of them showed up until the C++ leg of the
differential was run from a clean checkout.

durability_test.cpp added a uint8_t to a literal and initialized a uint64_t
field with the result. The addition promotes to int, and a braced initializer
will not narrow, so the file did not compile.

oraclevm compiles two of fhevm's sources directly and named them at a fixed
path beside itself. fhevm is in compute now, so that path is empty. It looks
beside itself first, then in compute, and takes -DFHEVM_DIR to be told
outright; a tree that still carries fhevm in place keeps working, and the
error now says what to pass when neither answers.

With these and the two CMake paths corrected in compute, the differential runs
942 vectors through Go, Rust and C++ and they agree on every field.

+++ chains/cpp/platformvm/test/durability_test.cpp
@@ -123,7 +123,7 @@ txs::Tx a_transaction(std::uint8_t seq) {
     b.network_id = 96369;
     b.blockchain_id = id_of(0x20);
     b.outs = {TransferableOutput{id_of(0x10), 0,
-                                 TransferOutput{1'000'000 + seq, OutputOwners{0, 1, {short_of(0x30)}}}}};
+                                 TransferOutput{static_cast<std::uint64_t>(1'000'000 + seq), OutputOwners{0, 1, {short_of(0x30)}}}}};
     b.ins = {TransferableInput{UtxoId{id_of(0x60), seq}, id_of(0x10), 0,
                                TransferInput{1'000'042, {0}}}};
     return signed_free(txs::BaseTxUnsigned::create(b).value());
